Simplifying fn functionality
The Cursor keys switch one key to the right to fit right under the
normal position of the left hand (ESDF).
Numlock is now only available in the fn layer.
fn-345 are Print, Scroll Lock and Pause.
fn-6 is ctr-break
fn-7 is win-pause
fn-b is LGUI instead of g to minimize accidental start menu when using
the cursors.